+++ nodejs/rest-api/rest-api-004/setup/docker/repo-mongo/setup-repo-mongo.sh
@@ -1,15 +1,14 @@
 cd $PROJECT_DIR/setup/docker/repo-mongo
 
 echo "################################################################################"
-echo " TIER REPO:${SERVER_ENV}"
+echo " setup-repo-mongo.sh:${SERVER_ENV}"
 echo "################################################################################"
-
-################################################
-# FOR DEEP CLEAN UP: REMOVE COMMENTS 
-################################################
-docker compose down;
-docker volume prune -f; 
-docker image prune -f; 
-
+    
+if [ "${DOCKER_DELETE_TIER_REPO_MONGO}" = true ]; then
+    echo "DELETING DOCKER TIER"
+    docker compose down;
+    docker volume prune -f; 
+    docker image prune -f; 
+fi
 
 docker compose up -d;
